TransformProfile.java rotation fix
This file contained two bugs. First, the coordinate transformation logic in the transform methods was flawed, incorrectly applying coordinate swaps for 180-degree rotations. I've rewritten this logic to be clearer and correct for all rotation angles. Second, the rotate(Direction) method had a typo (case 190) which I've corrected to case 180.